+++ packages/punchcard/lib/core/index.ts
@@ -2,7 +2,9 @@ export * from './app';
 export * from './assembly';
 export * from './cache';
 export * from './client';
+export * from './code';
 export * from './dependency';
 export * from './entrypoint';
 export * from './file-system';
+export * from './global';
 export * from './resource';
